Move all recursion-related code to a separate file/package
