+++ modules/generic/src/main/scala/ShowModule.scala
@@ -4,8 +4,9 @@ package schema
 
 package generic
 
+import recursion._
+
 trait ShowModule[R <: Realisation] extends GenericSchemaModule[R] {
-  import SchemaF._
 
   implicit val showDecidableInstance: Decidable[Show] = new Decidable[Show] {
     override def choose2[Z, A1, A2](a1: => Show[A1], a2: => Show[A2])(f: Z => A1 \/ A2): Show[Z] =
